refactor(db-interface): extracts helper funcions away from the db class
